Add the controller.

+++ EmployeePayslipCalculator/EmployeePayslipCalculator.Models/EmployeeInfo.cs
@@ -8,6 +8,6 @@ namespace EmployeePayslipCalculator.Models
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public int AnnualSalary { get; set; }
-        public decimal SuperRate { get; set; }
+        public double SuperRate { get; set; }
     }
 }
